# backend/openapi/horse-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Horse Reservation API",
    "description": "마장 예약 서비스 API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/admin/timeslots/{timeSlotId}/reservations/{reservationId}/cancel-for-closure": {
      "post": {
        "tags": [
          "admin-time-slot-closure-controller"
        ],
        "operationId": "cancelReservation",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleClosureCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCancelResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/timeslots/{timeSlotId}/closure": {
      "post": {
        "tags": [
          "admin-time-slot-closure-controller"
        ],
        "operationId": "start",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSlotClosureStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotClosureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/timeslots/{timeSlotId}/closure/withdraw": {
      "post": {
        "tags": [
          "admin-time-slot-closure-controller"
        ],
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSlotClosureActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotClosureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/timeslots/{timeSlotId}/closure/reopen": {
      "post": {
        "tags": [
          "admin-time-slot-closure-controller"
        ],
        "operationId": "reopen",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSlotClosureActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotClosureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/timeslots/{timeSlotId}/closure/complete": {
      "post": {
        "tags": [
          "admin-time-slot-closure-controller"
        ],
        "operationId": "complete",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSlotClosureActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotClosureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/schedule-dates/{scheduleDate}/reservations/{reservationId}/cancel-for-closure": {
      "post": {
        "tags": [
          "admin-schedule-date-controller"
        ],
        "operationId": "cancelReservation_1",
        "parameters": [
          {
            "name": "scheduleDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleClosureCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCancelResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/reservations/{reservationId}/complete": {
      "post": {
        "tags": [
          "admin-reservation-completion-controller"
        ],
        "operationId": "complete_1",
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationCompletionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/timeslots/{timeSlotId}": {
      "delete": {
        "tags": [
          "admin-time-slot-controller"
        ],
        "operationId": "deleteTimeSlot",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "admin-time-slot-controller"
        ],
        "operationId": "changeClosedStatus",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSlotStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/timeslots/{timeSlotId}/closure-impact": {
      "get": {
        "tags": [
          "admin-time-slot-closure-controller"
        ],
        "operationId": "getClosure",
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotClosureResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TimeSlotResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "lessonDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string"
          },
          "totalCapacity": {
            "type": "integer",
            "format": "int32"
          },
          "roundArenaCapacity": {
            "type": "integer",
            "format": "int32"
          },
          "classCapacities": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "closed": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReservationCancelResponse": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "responsibility": {
            "type": "string"
          },
          "couponAction": {
            "type": "string"
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time"
          },
          "changed": {
            "type": "boolean"
          }
        }
      },
...
      "ScheduleClosureCancelRequest": {
        "type": "object",
        "properties": {
          "memo": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          }
        },
        "required": [
          "memo"
        ]
      },
      "TimeSlotClosureStartRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          }
        },
        "required": [
          "reason"
        ]
      },
      "TimeSlotClosureImpactResponse": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "integer",
            "format": "int64"
          },
          "reservationStatusAtStart": {
            "type": "string"
          },
          "currentStatus": {
            "type": "string"
          },
          "resolved": {
            "type": "boolean"
          },
          "moved": {
            "type": "boolean"
          },
          "memberId": {
            "type": "integer",
            "format": "int64"
          },
          "memberName": {
            "type": "string"
          },
          "memberPhone": {
            "type": "string"
          },
          "classType": {
            "type": "string"
          },
          "paymentSource": {
            "type": "string"
          },
          "couponId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "classType",
          "currentStatus",
          "memberId",
          "memberName",
          "memberPhone",
          "moved",
          "paymentSource",
          "reservationId",
          "reservationStatusAtStart",
          "resolved"
        ]
      },
      "TimeSlotClosureResponse": {
        "type": "object",
        "properties": {
          "timeSlotId": {
            "type": "integer",
            "format": "int64"
          },
          "adminClosed": {
            "type": "boolean"
          },
          "closed": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_PROGRESS",
              "COMPLETED",
              "WITHDRAWN"
            ]
          },
          "reason": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "withdrawnAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "resolvedCount": {
            "type": "integer",
            "format": "int32"
          },
          "unresolvedCount": {
            "type": "integer",
            "format": "int32"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "impacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeSlotClosureImpactResponse"
            }
          }
        },
        "required": [
          "adminClosed",
          "closed",
          "impacts",
          "progressPercent",
          "reason",
          "resolvedCount",
          "startedAt",
          "status",
          "timeSlotId",
          "totalCount",
          "unresolvedCount",
          "version"
        ]
      },
      "TimeSlotClosureActionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "expectedVersion",
          "reason"
        ]
      },
...
      "ReservationCompletionResponse": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "paymentSource": {
            "type": "string"
          },
          "couponId": {
            "type": "integer",
            "format": "int64"
          },
          "generalRideCount": {
            "type": "integer",
            "format": "int32"
          },
          "dressageRideCount": {
            "type": "integer",
            "format": "int32"
          },
          "jumpingRideCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "TimeSlotStatusUpdateRequest": {
        "type": "object",
        "properties": {
          "closed": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "details",
          "fieldErrors",
          "message",
          "path",
          "status",
          "timestamp"
        ]
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}